# src/swagger_doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Starnet API",
    "description": ""
  },
  "servers": [
    {
      "url": "https://dev.starnet.ai/api.staging",
      "description": ""
    }
  ],
  "basePath": "",
...
  "paths": {
...
    "/brands/sign-up": {
      "post": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "country_code": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "passwd": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/brands/campaign/applicants": {
      "get": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "query",
            "description": "the Campaign ID to query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "the status of the applicant, default be 'open'",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/campaignApplicants"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "current page, default 1",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "items dispayed per page, default 10",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/brands/campaign/applicant/approve": {
      "post": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "any"
                  },
                  "campaign_id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/brands/campaign/applicant/reject": {
      "post": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "any"
                  },
                  "campaign_id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/brands/campaign/post/review": {
      "patch": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "any"
                  },
                  "campaign_id": {
                    "example": "any"
                  },
                  "comment_photo": {
                    "example": "any"
                  },
                  "comment_text": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/brands/campaign/post/approve": {
      "patch": {
        "tags": [
          "Brand"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "example": "any"
                  },
                  "campaign_id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications/unread": {
      "get": {
        "tags": [
          "Notification"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/notifications/read": {
      "post": {
        "tags": [
          "Notification"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note_ids": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/test/{user_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "campaignApplicants": {
        "type": "string",
        "enum": [
          "open",
          "applied",
          "declined",
          "approved",
          "post-draft",
          "post-submitted",
          "post-launched"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}